current: move dataSource out of FluMap
Since the SeasonTimeline will need the same dataSource to aggregate
the overall flu conditions, move the dataSource state up to the parent
component, CurrentConditions.
